+++ src/config.rs
@@ -4,7 +4,7 @@ use std::path::PathBuf;
 use defaults::Defaults;
 use serde::Deserialize;
 
-use crate::style::{Margin, Padding};
+use crate::style::{Margin, Padding, Radius};
 use crate::Color;
 
 const DEFAULT_CONFIG_NAME: &str = concat!(crate::prog_name!(), ".config");
@@ -35,6 +35,8 @@ pub struct Config {
     font_size: Option<u16>,
     bg_color: Option<Color>,
     font_color: Option<Color>,
+    #[def = "Radius::all(0.0)"]
+    corner_radius: Radius,
 
     icon: Option<Icon>,
 
@@ -70,6 +72,8 @@ struct InputText {
     margin: Margin,
     #[def = "Padding::from_pair(1.7, -4.0)"]
     padding: Padding,
+    #[def = "Radius::all(f32::MAX)"]
+    corner_radius: Radius,
 }
 
 #[derive(Defaults, Deserialize)]