+++ src/style.rs
@@ -22,6 +22,14 @@ pub struct Margin {
     pub right: f32,
 }
 
+#[derive(Clone, Default)]
+pub struct Radius {
+    pub top_left: f32,
+    pub top_right: f32,
+    pub bottom_left: f32,
+    pub bottom_right: f32,
+}
+
 impl Padding {
     pub const fn all(val: f32) -> Self {
         Self {
@@ -75,6 +83,48 @@ impl Margin {
     }
 }
 
+impl Mul<f32> for &Radius {
+    type Output = Radius;
+
+    fn mul(self, rhs: f32) -> Radius {
+        Radius {
+            top_left: self.top_left * rhs,
+            top_right: self.top_right * rhs,
+            bottom_left: self.bottom_left * rhs,
+            bottom_right: self.bottom_right * rhs,
+        }
+    }
+}
+
+impl Radius {
+    pub const fn all(val: f32) -> Self {
+        Self {
+            top_left: val,
+            top_right: val,
+            bottom_left: val,
+            bottom_right: val,
+        }
+    }
+
+    pub const fn from_pair(first: f32, second: f32) -> Self {
+        Self {
+            top_left: first,
+            top_right: second,
+            bottom_left: second,
+            bottom_right: first,
+        }
+    }
+
+    pub(crate) fn min(&self, other: Radius) -> Radius {
+        Self {
+            top_left: self.top_left.min(other.top_left),
+            top_right: self.top_right.min(other.top_right),
+            bottom_left: self.bottom_left.min(other.bottom_left),
+            bottom_right: self.bottom_right.min(other.bottom_right),
+        }
+    }
+}
+
 impl Mul<f32> for &Margin {
     type Output = Margin;
 
@@ -106,6 +156,15 @@ impl<'de> Deserialize<'de> for Margin {
     }
 }
 
+impl<'de> Deserialize<'de> for Radius {
+    fn deserialize<D>(d: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        d.deserialize_str(StringVisitor(PhantomData))
+    }
+}
+
 struct StringVisitor<T>(PhantomData<T>);
 
 impl<'de, T> Visitor<'de> for StringVisitor<T>
@@ -127,16 +186,38 @@ where
     }
 }
 
-impl FromStr for Padding {
-    type Err = &'static str;
+struct FiniteFloatVec {
+    pub values: Vec<f32>,
+}
+
+impl FromStr for FiniteFloatVec {
+    type Err = String;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         let values = s
             .split(' ')
             .map(|s| s.trim())
             .filter(|s| !s.is_empty())
-            .map(|s| s.parse::<f32>().map_err(|_| "invalid float value"))
+            .map(|s| {
+                s.parse::<f32>()
+                    .map_err(|_| format!("invalid float value: {:?}", s))
+                    .and_then(|f| {
+                        f.is_finite()
+                            .then_some(f)
+                            .ok_or(format!("non-finite float value: {:?}", f))
+                    })
+            })
             .collect::<Result<Vec<f32>, _>>()?;
+
+        Ok(Self { values })
+    }
+}
+
+impl FromStr for Padding {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let values = FiniteFloatVec::from_str(s)?.values;
 
         match values.len() {
             1 => Ok(Self::all(values[0])),
@@ -147,21 +228,16 @@ impl FromStr for Padding {
                 left: values[3],
                 right: values[1],
             }),
-            _ => Err("padding should consists of either 1, 2 or 4 floats"),
+            _ => Err("padding should consists of either 1, 2 or 4 floats".into()),
         }
     }
 }
 
 impl FromStr for Margin {
-    type Err = &'static str;
+    type Err = String;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let values = s
-            .split(' ')
-            .map(|s| s.trim())
-            .filter(|s| !s.is_empty())
-            .map(|s| s.parse::<f32>().map_err(|_| "invalid float value"))
-            .collect::<Result<Vec<f32>, _>>()?;
+        let values = FiniteFloatVec::from_str(s)?.values;
 
         match values.len() {
             1 => Ok(Self::all(values[0])),
@@ -172,7 +248,35 @@ impl FromStr for Margin {
                 left: values[3],
                 right: values[1],
             }),
-            _ => Err("margin should consists of either 1, 2 or 4 floats"),
+            _ => Err("margin should consists of either 1, 2 or 4 floats".into()),
+        }
+    }
+}
+
+impl FromStr for Radius {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let values = FiniteFloatVec::from_str(s)?
+            .values
+            .iter()
+            .map(|&f| {
+                (!f.is_sign_negative())
+                    .then_some(f)
+                    .ok_or(format!("radius can not be negative: {:?}", f))
+            })
+            .collect::<Result<Vec<f32>, _>>()?;
+
+        match values.len() {
+            1 => Ok(Self::all(values[0])),
+            2 => Ok(Self::from_pair(values[0], values[1])),
+            4 => Ok(Self {
+                top_left: values[0],
+                top_right: values[1],
+                bottom_left: values[3],
+                bottom_right: values[2],
+            }),
+            _ => Err("radius should consists of either 1, 2 or 4 floats".into()),
         }
     }
 }